Delete cart sub-sections and refactor cart UI
Removed 'main-cart-footer' and 'main-cart-items' sections and refactored the structure of the cart.JSON to use 'cart' section instead. Also applied significant improvements to the user interface of the cart in 'cart.liquid' section, presenting a more streamlined and easy to navigate layout for the users. Minor modifications were also made to the product title in 'main-product.liquid'.